update soft delete
